Solution to Check Greater Than or Equal To

+++ Check Greater Than or Equal To@@67882a731b4392d8fc1577d8/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    int a,b;
+    float a,b;
     scanf("%f %f",&a,&b);
     ((a>b) || (a==b))?printf("True"):printf("False");
     return 0;
